Refactor autocomplete network
